Implemented new routing system
